chain/state: Refactor in preparation for adding hooks
Preparatory work for adding support for initializing the `last_sign`
state from a user-specified hook.
